third_party/WebKit/LayoutTests/inspector-protocol/layout-fonts/ogham.html

<!DOCTYPE html>
<html>
<meta charset="UTF-8">
<head>
  <script type="text/javascript" src="resources/inspector-protocol-test.js"></script>
  <script type="text/javascript" src="resources/css-protocol-test.js"></script>
  <script type="text/javascript" src="resources/dom-protocol-test.js"></script>
    <script type="text/javascript" src="resources/layout-font-test.js"></script>
</head>
<body>
  <!-- All ogham glyps in two font families. -->
  <div class="test">
    <div id=oghammonofont></div>
    <div id=oghamdefaultfont></div>
    <script>
      function makeCharHtml(charCode) {
        return "&#x" + charCode.toString(16) + ";";
      }
      function makeCharHtmlSequence(first, last) {
        var result = "";
        for (var i = first; i <= last; i++)
          result += makeCharHtml(i);
        return result;
      }
      charSequenceHtml = makeCharHtmlSequence(0x1680, 0x169c);
      document.getElementById("oghammonofont").innerHTML = charSequenceHtml;
      document.getElementById("oghamdefaultfont").innerHTML = charSequenceHtml;
    </script>
  </div>
  There should be two lines of
  <a href="https://en.wikipedia.org/wiki/Ogham">Ogham</a> above.
</body>
</html>
