third_party/WebKit/LayoutTests/platform/win-xp/inspector-protocol/layout-fonts/ogham-expected.txt

 ᚁᚂᚃᚄᚅᚆᚇᚈᚉᚊᚋᚌᚍᚎᚏᚐᚑᚒᚓᚔᚕᚖᚗᚘᚙᚚ᚛᚜
#oghammonofont:
"Arial" : 29

 ᚁᚂᚃᚄᚅᚆᚇᚈᚉᚊᚋᚌᚍᚎᚏᚐᚑᚒᚓᚔᚕᚖᚗᚘᚙᚚ᚛᚜
#oghamdefaultfont:
"Arial" : 29

There should be two lines of Ogham above.
